+++ WebApp/settings.py
@@ -20,7 +20,7 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 # See https://docs.djangoproject.com/en/3.1/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = ''
+SECRET_KEY = '2@v$fyqv10xmb+@@*%0($12cg21$uu5jteiidwoklqj@k@@+bk'
 
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
