+++ debug.log
@@ -1,0 +1,1 @@
+[0326/181027.911:ERROR:registration_protocol_win.cc(130)] TransactNamedPipe: The pipe has been ended. (0x6D)
