+++ drugs/example_input.csv
@@ -1,0 +1,2 @@
+canonical_smiles
+CCc1ncnc(-c2cc(F)c(C(=O)N3CCN4CCC[C@H]4C3)c(Cl)c2)c1C#Cc1ccc(N)nc1
